:package: 마이페이지 관련 코드 패키지 분리
