+++ presentation/src/main/java/com/whyranoid/presentation/screens/mypage/tabs/ChallengePage.kt
@@ -1,0 +1,7 @@
+package com.whyranoid.presentation.screens.mypage.tabs
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun ChallengePage() {
+}
